requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.20
spotipy==2.23.0
python-dotenv==1.0.0
Flask-Session==0.6.0
aiohttp==3.9.3
httpx==0.23.0
cachetools==5.3.2
ultralytics==8.0.198
opencv-python==4.8.1.78